+++ examples/provider/provider.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     zabbix = {
-      source = "adi/zabbix"
+      source = "Tensai123/zabbix"
     }
   }
 }